==> trifeature/04b-scikitlearn-improve.ipynb <==
# Try to improve LinearDiscriminantAnalysis

TLDR: Settings don't matter. Use defaults.

## Load Stuff

import numpy as np
import pandas as pd
import time
import sys
import joblib

import sklearn
from sklearn import *
from sklearn.experimental import enable_hist_gradient_boosting
import skopt

sys.path.append("..")
from helpers import filename_for
from plotconf import *

files = [filename_for(15, 30, 600, 500, n, "inclxx", s, "trifeature.pkl") for n in [1, 2, 3, 4] for s in range(20)]
dfs = [pd.read_pickle(file) for file in files]
data = pd.concat(dfs, ignore_index=True).sample(frac=1)
data = data[data["nHits"] > 0]
print(data.shape)

msk = np.random.rand(len(data)) < 0.8
traindata = data[msk]
testdata = data[~msk]

print(traindata.shape)
print(testdata.shape)

features = ["nHits", "nClus", "Edep"]
label = ["nPN"]

## Try to improve ...

def optimize_lda1():
    defaults = {
        "solver": "svd",
        "shrinkage": None,
        "priors": None,
        "n_components": None,
        "store_covariance": False,
        "tol": 0.0001,
    }

    model = sklearn.discriminant_analysis.LinearDiscriminantAnalysis(**defaults)

    opt = skopt.BayesSearchCV(
        model,
        {
            # "solver": skopt.space.Categorical(["svd", "lsqr", "eigen"]),
            # "shrinkage": skopt.space.Categorical(["auto", None]), # Works only with ‘lsqr’ and ‘eigen’ solvers.
            "n_components": skopt.space.Categorical([1, 2, 3]),
            "tol": skopt.space.Real(1e-3, 1e0, prior="log-uniform"),  # Only used if solver is ‘svd’.
        },
        n_iter=50,
        cv=2,
        n_jobs=2,
    )

    start = time.time()
    opt.fit(traindata[features], traindata[label].values.ravel())
    end = time.time()

    y_pred = opt.predict(testdata[features])
    y_true = testdata[label].values.ravel()
    bac = sklearn.metrics.balanced_accuracy_score(y_true, y_pred)

    y_pred = opt.predict(data[features])
    y_true = data[label].values.ravel()
    bacall = sklearn.metrics.balanced_accuracy_score(y_true, y_pred)

    return opt, ("LinearDiscriminantAnalysis", end - start, bac, bacall)

%%capture
lda1_opt, lda1_result = optimize_lda1()

print(lda1_result)
print(lda1_opt.best_params_)
display(pd.DataFrame(lda1_opt.cv_results_))

def optimize_lda2():
    defaults = {
        "solver": "svd",
        "shrinkage": None,
        "priors": None,
        "n_components": None,
        "store_covariance": False,
        "tol": 0.0001,
    }

    model = sklearn.discriminant_analysis.LinearDiscriminantAnalysis(**defaults)

    opt = skopt.BayesSearchCV(
        model,
        {
            "solver": skopt.space.Categorical(["lsqr", "eigen"]),
            # "shrinkage": skopt.space.Categorical(["auto"]),#, None]),  # Works only with ‘lsqr’ and ‘eigen’ solvers.
            "n_components": skopt.space.Categorical([1, 2, 3]),
            # "tol": skopt.space.Real(1e-3, 1e0, prior="log-uniform"), # Only used if solver is ‘svd’.
        },
        n_iter=50,
        cv=2,
        n_jobs=2,
    )

    start = time.time()
    opt.fit(traindata[features], traindata[label].values.ravel())
    end = time.time()

    y_pred = opt.predict(testdata[features])
    y_true = testdata[label].values.ravel()
    bac = sklearn.metrics.balanced_accuracy_score(y_true, y_pred)

    y_pred = opt.predict(data[features])
    y_true = data[label].values.ravel()
    bacall = sklearn.metrics.balanced_accuracy_score(y_true, y_pred)

    return opt, ("LinearDiscriminantAnalysis", end - start, bac, bacall)

%%capture
lda2_opt, lda2_result = optimize_lda2()

print(lda2_result)
print(lda2_opt.best_params_)
display(pd.DataFrame(lda2_opt.cv_results_))

## Just train model with defaults settings and save

model = sklearn.discriminant_analysis.LinearDiscriminantAnalysis()
model.fit(traindata[features], traindata[label].values.ravel())

y_pred = model.predict(testdata[features])
print(sklearn.metrics.balanced_accuracy_score(testdata[label], y_pred))

joblib.dump(model, "models/mult_30dp_600AMeV_4n_LinearDiscriminantAnalysis.pkl")
